🍌 Add[dev]: drawLineLoop
- To use GL_LINE_LOOP to draw the lines
Where: 3D_tools.cpp/.h

+++ include/corridor.h
@@ -20,7 +20,7 @@ class Corridor
 
     private:
         int m_x;
-        int m_y;
+        float m_y;
         int m_z;
         int m_start;
         float m_walk;
